Prepare for review
- split rust native crate for modules
- add Rc for video source
- remove "0" call
- add ids` types
