+++ crates/libwebrtc-sys/cpp_src/device_video_capturer.cpp
