+++ crates/native/Cargo.toml
@@ -10,7 +10,6 @@ crate-type = ["cdylib"]
 [dependencies]
 libwebrtc-sys = { path = "../libwebrtc-sys" }
 cxx = "1.0"
-lazy_static = "1.4.0"
 
 [build-dependencies]
 anyhow = "1.0"
